Actual color burn mix-blend-mode SVGs
Okay these are exported using a color burn mix-blend-mode in the SVG. That said, couldn't do as much out-of-illustrator enhancements. Hoping the asset export worked without borking the bounding box, but these files look so much better and are about half the size — they check out in-browser so fuck it. Can revert if needed