PHEW - modal event routing
This was like looking for a specific piece of spaghetti.

also established a concept for echoed events — things that serve notice in the main window as well as their embed. these allow both windows to react to the same event, separately or coordinated.

right now this is event driven — so no callback. you can add a queryName to the modal request to use in catching the event and differentiating between multiple modal prompts. i think there's room to add callbacks later with this as the underlying structure...but that's future enhancement stuff.